requirements.txt
aiohttp==3.8.1
aiosignal==1.2.0
async-timeout==4.0.2
attrs==21.4.0
autopep8==1.6.0
charset-normalizer==2.0.12
colorlog==6.6.0
frozenlist==1.3.0
hikari==2.0.0.dev108
hikari-lightbulb==2.2.1
idna==3.3
motor==3.0.0
multidict==6.0.2
pycodestyle==2.8.0
pymongo==4.1.1
toml==0.10.2
yarl==1.7.2
